First stab at falling block handling
There is an issue with falling blocks inside water or lava because they
are slowed. Might need to put in a sleep timer to make the turtle wait a
few ticks for the block to fall a bit more.  This also seems to wreck
the column counter which throws the whole system off.